+++ yandex/resource_yandex_compute_instance.go
@@ -509,6 +509,25 @@ func resourceYandexComputeInstance() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+
+			"local_disk": {
+				Type:         schema.TypeList,
+				Optional:     true,
+				RequiredWith: []string{"placement_policy"},
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"size_bytes": {
+							Type:     schema.TypeInt,
+							Required: true,
+							ForceNew: true,
+						},
+						"device_name": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+					},
+				},
+			},
 		},
 	}
 }
@@ -598,6 +617,8 @@ func resourceYandexComputeInstanceRead(d *schema.ResourceData, meta interface{})
 		return err
 	}
 
+	localDisks := flattenLocalDisks(instance)
+
 	d.Set("created_at", getTimestamp(instance.CreatedAt))
 	d.Set("platform_id", instance.PlatformId)
 	d.Set("folder_id", instance.FolderId)
@@ -639,6 +660,10 @@ func resourceYandexComputeInstanceRead(d *schema.ResourceData, meta interface{})
 	}
 
 	if err := d.Set("placement_policy", placementPolicy); err != nil {
+		return err
+	}
+
+	if err := d.Set("local_disk", localDisks); err != nil {
 		return err
 	}
 
@@ -1233,6 +1258,8 @@ func prepareCreateInstanceRequest(d *schema.ResourceData, meta *Config) (*comput
 		return nil, fmt.Errorf("Error create 'placement_policy' object of api request: %s", err)
 	}
 
+	localDisks := expandLocalDiskSpecs(d.Get("local_disk"))
+
 	req := &compute.CreateInstanceRequest{
 		FolderId:              folderID,
 		Hostname:              d.Get("hostname").(string),
@@ -1250,6 +1277,7 @@ func prepareCreateInstanceRequest(d *schema.ResourceData, meta *Config) (*comput
 		NetworkInterfaceSpecs: nicSpecs,
 		SchedulingPolicy:      schedulingPolicy,
 		PlacementPolicy:       placementPolicy,
+		LocalDiskSpecs:        localDisks,
 	}
 
 	return req, nil