+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 * alb: `send` and `receive` attributes can be empty now in `yandex_alb_backend_group` resource and data source
 * alb: fixed `yandex_alb_load_balancer` resource and data source documentation
 
+FEATURES:
+* compute: support for creating `yandex_compute_instance` with local disks
+
 ## 0.75.0 (May 31, 2022)
 ENHANCEMENTS:
 * compute: support instance's scheduling policy update in-place